Fix warning "[repositoryPath] is not valid git repository for [repositoryLocation], will fetch again"
SPM will produce this warning while creating or updating the workspace if the remote origin URL of the repo does not match specified value. It will then discard the workspace clone and copy the repo from cache, or clone again if cache is not available.
This issue arises because SPM does not normalize/canonicalize URLs that git treats as equivalent, e.g. "https://github.com/swiftlang/swift-package-manager" and "https://github.com/swiftlang/swift-package-manager.git" (and arguably, the SSH versions and maybe other). If the origin URL specified in the cached repo differs from the URL in the dependency graph, the workspace operation will succeed, but subsequent operations will trigger the issue again, leading to slower workspace operations and increased disk activity.